create reusable ProjectsFiltersButton component
